+++ filament/backend/src/opengl/OpenGLContext.cpp
@@ -60,21 +60,28 @@ OpenGLContext::OpenGLContext() noexcept {
     glGetIntegerv(GL_MAX_SAMPLES, &gets.max_samples);
     glGetIntegerv(GL_MAX_DRAW_BUFFERS, &gets.max_draw_buffers);
     glGetIntegerv(GL_MAX_TEXTURE_IMAGE_UNITS, &gets.max_texture_image_units);
+    glGetIntegerv(GL_MAX_COMBINED_TEXTURE_IMAGE_UNITS, &gets.max_combined_texture_image_units);
     glGetIntegerv(GL_UNIFORM_BUFFER_OFFSET_ALIGNMENT, &gets.uniform_buffer_offset_alignment);
+
+    constexpr auto const caps3 = FEATURE_LEVEL_CAPS[+FeatureLevel::FEATURE_LEVEL_3];
+    constexpr GLint MAX_VERTEX_SAMPLER_COUNT = caps3.MAX_VERTEX_SAMPLER_COUNT;
+    constexpr GLint MAX_FRAGMENT_SAMPLER_COUNT = caps3.MAX_FRAGMENT_SAMPLER_COUNT;
 
     if constexpr (BACKEND_OPENGL_VERSION == BACKEND_OPENGL_VERSION_GLES) {
         initExtensionsGLES();
         if (state.major == 3) {
             assert_invariant(gets.max_texture_image_units >= 16);
-            if (state.minor >= 0) {
-                mShaderModel = ShaderModel::MOBILE;
-            }
+            assert_invariant(gets.max_combined_texture_image_units >= 32);
             if (state.minor >= 1) {
                 features.multisample_texture = true;
                 // figure out our feature level
-                if (gets.max_texture_image_units >= 31 &&
-                    ext.EXT_texture_cube_map_array) {
+                if (ext.EXT_texture_cube_map_array) {
                     mFeatureLevel = FeatureLevel::FEATURE_LEVEL_2;
+                    if (gets.max_texture_image_units >= MAX_FRAGMENT_SAMPLER_COUNT &&
+                        gets.max_combined_texture_image_units >=
+                                (MAX_FRAGMENT_SAMPLER_COUNT + MAX_VERTEX_SAMPLER_COUNT)) {
+                        mFeatureLevel = FeatureLevel::FEATURE_LEVEL_3;
+                    }
                 }
             }
         }
@@ -82,14 +89,16 @@ OpenGLContext::OpenGLContext() noexcept {
         // OpenGL version
         initExtensionsGL();
         if (state.major == 4) {
-            if (state.minor >= 1) {
-                mShaderModel = ShaderModel::DESKTOP;
-            }
+            assert_invariant(state.minor >= 1);
+            mShaderModel = ShaderModel::DESKTOP;
             if (state.minor >= 3) {
+                // cubemap arrays are available as of OpenGL 4.0
+                mFeatureLevel = FeatureLevel::FEATURE_LEVEL_2;
                 // figure out our feature level
-                if (gets.max_texture_image_units >= 31) {
-                    // cubemap arrays are available as of OpenGL 4.0
-                    mFeatureLevel = FeatureLevel::FEATURE_LEVEL_2;
+                if (gets.max_texture_image_units >= MAX_FRAGMENT_SAMPLER_COUNT &&
+                    gets.max_combined_texture_image_units >=
+                            (MAX_FRAGMENT_SAMPLER_COUNT + MAX_VERTEX_SAMPLER_COUNT)) {
+                    mFeatureLevel = FeatureLevel::FEATURE_LEVEL_3;
                 }
             }
             features.multisample_texture = true;
@@ -97,8 +106,8 @@ OpenGLContext::OpenGLContext() noexcept {
         // feedback loops are allowed on GL desktop as long as writes are disabled
         bugs.allow_read_only_ancillary_feedback_loop = true;
         assert_invariant(gets.max_texture_image_units >= 16);
+        assert_invariant(gets.max_combined_texture_image_units >= 32);
     }
-
 #ifdef GL_EXT_texture_filter_anisotropic
     if (ext.EXT_texture_filter_anisotropic) {
         glGetFloatv(GL_MAX_TEXTURE_MAX_ANISOTROPY_EXT, &gets.max_anisotropy);
@@ -213,6 +222,7 @@ OpenGLContext::OpenGLContext() noexcept {
         // TODO: see if we could use `bugs.allow_read_only_ancillary_feedback_loop = true`
     }
 
+    slog.v << "Feature level: " << +mFeatureLevel << '\n';
     slog.v << "Active workarounds: " << '\n';
     UTILS_NOUNROLL
     for (auto [enabled, name, _] : mBugDatabase) {