+++ filament/backend/src/opengl/OpenGLContext.h
@@ -28,6 +28,7 @@
 
 #include <array>
 #include <set>
+#include <tuple>
 #include <utility>
 
 namespace filament::backend {
@@ -35,6 +36,7 @@ namespace filament::backend {
 class OpenGLContext {
 public:
     static constexpr const size_t MAX_TEXTURE_UNIT_COUNT = MAX_SAMPLER_COUNT;
+    static constexpr const size_t DUMMY_TEXTURE_BINDING = 31; // highest binding less than 32
     static constexpr const size_t MAX_BUFFER_BINDINGS = 32;
     typedef math::details::TVec4<GLint> vec4gli;
     typedef math::details::TVec2<GLclampf> vec2glf;
@@ -126,6 +128,7 @@ public:
         GLint max_samples;
         GLint max_uniform_block_size;
         GLint max_texture_image_units;
+        GLint max_combined_texture_image_units;
         GLint uniform_buffer_offset_alignment;
     } gets = {};